Extracted bot utility methods
